+++ src/main/java/com/example/demo/jpa/QuizDetailRepository.java
@@ -4,5 +4,5 @@ import com.example.demo.entity.QuizDetailEntity;
 import com.example.demo.entity.QuizEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface QuizDetailRepository extends JpaRepository<QuizDetailEntity, Integer> {
+public interface QuizDetailRepository extends JpaRepository<QuizDetailEntity, Long> {
 }
